style: Extracted the interfaces and types out of the context files and moved them to the types folder.
