+++ src/types/context/customization.types.ts
@@ -1,0 +1,9 @@
+export type Selection = {
+  [customizationId: string]: string;
+};
+
+export interface CustomizationContextType {
+  selections: Selection;
+  handleSelection: (customizationId: string, optionId: string) => void;
+  clearSelections: () => void;
+}
